Separate `DnsSettings` and `StaticResolvConf`
